Revert "Moving to new directories (#38)"
This reverts commit 347472754b7fdb8b205146b9c123e20e0b1e936d.
