+++ ahorratron/sync_api/institutions/banco_de_chile/demo.py
@@ -1,7 +1,7 @@
 import json
 from pathlib import Path
 
-from .models import (
+from ahorratron.sync_api.institutions.banco_de_chile.models import (
     GetCartolaCuentaRequest,
     GetCartolaResponse,
     GetSaldoResponse,
