+++ ahorratron/sync_api/models/core_models.py
@@ -1,3 +1,5 @@
+from typing import Optional
+
 from pydantic import BaseModel
