+++ docker-compose.yml
@@ -51,7 +51,7 @@ services:
             -days 365 \
             -subj "/CN=api.pluggy.ai";
         fi;
-        uv run uvicorn sync_api.main:app --host 0.0.0.0 --port 443 --ssl-keyfile /certs/key.pem --ssl-certfile /certs/cert.pem
+        uv run uvicorn ahorratron.sync_api.main:app --host 0.0.0.0 --port 443 --ssl-keyfile /certs/key.pem --ssl-certfile /certs/cert.pem
       '
     networks:
       ahorratron_net:
