+++ pyproject.toml
@@ -19,9 +19,9 @@ dependencies = [
 ]
 
 [project.scripts]
-bank-statement = "conversor.cli.bank_statement:main"
-convert-to-actual = "conversor.cli.convert_to_actual:main"
-actual-api = "actual_api.main:run_server"
+bank-statement = "ahorratron.cli.bank_statement:main"
+convert-to-actual = "ahorratron.cli.convert_to_actual:main"
+ahorratron-api = "ahorratron.api.main:run_server"
 
 [dependency-groups]
 dev = [
@@ -33,7 +33,7 @@ dev = [
 ]
 
 [tool.setuptools]
-packages = ["ahorratron", "conversor", "sync_api", "actual_api"]
+packages = ["ahorratron"]
 
 [tool.uv]
 package = true
